ISSUE-151: Poorman's async
